+++ app/services/art_service/patient_state_engine.rb
@@ -5,6 +5,8 @@ module ARTService
   class PatientStateEngine
     attr_accessor :patient, :date
 
+    include ModelUtils
+
     def initialize(patient, date)
       @patient = patient
       @date = date
@@ -57,7 +59,7 @@ module ARTService
                      .exists?
     end
 
-    def mark_patient_art_start_date
+    def mark_patient_art_start_date(patient)
       art_start_date_concept = concept('ART start date')
       has_art_start_date = Observation.where(person_id: patient.patient_id,
                                              concept: art_start_date_concept)