+++ app/services/report_service.rb
@@ -169,6 +169,10 @@ class ReportService
     engine(@program).patient_visit_list(start_date, end_date)
   end
 
+  def patient_outcome_list(start_date, end_date, outcome)
+    engine(@program).patient_outcome_list(start_date, end_date, outcome)
+  end
+
   private
 
   def engine(program)